+++ src/org/pvemu/game/objects/dep/Creature.java
@@ -20,7 +20,7 @@ public interface Creature {
      *
      * @return
      */
-    public short getInitiative();
+    public int getInitiative();
 
     public short getLevel();
